refac: divide round action buttons
