+++ book-talk-fe/src/routes/Debate/DebatePage.tsx
@@ -7,8 +7,8 @@ import {DebateHeader} from './_components/DebateHeader';
 import {DebatePresentation} from './_components/DebatePresentation';
 import {DebateMemberList} from './_components/DebateMemberList';
 import {RoundStartBackdropContainer} from './_components/RoundStartBackdropContainer';
-import {MicrophoneControl} from './_components/MicrophoneControl';
-import {ActionButton, DebateContainer} from './Debate.style';
+import {RoundActions} from './_components/RoundActions';
+import {DebateContainer} from './Debate.style';
 import {useDebate} from "../../hooks/useDebate.tsx";
 import {useDebateWebSocket} from "../../hooks/useDebateWebSocket.tsx";
 import {useDebateRoundSpeaker} from "../../hooks/useDebateRoundSpeaker";
@@ -16,71 +16,11 @@ import StartDebateModal from "./_components/StartDebateModal.tsx";
 import {VoiceChatProvider} from "../../contexts/VoiceChatContext";
 import type {WebSocketMessage} from "../../apis/websocket";
 import {findOneDebateQueryOptions} from "../../apis/debate";
-import raiseHandSvg from "../../assets/raise-hand.svg";
 
 type RoundType = 'PREPARATION' | 'PRESENTATION' | 'FREE';
 
 interface Props {
     debateId: string | undefined;
-}
-
-interface RoundActionsProps {
-    roundType: RoundType;
-    myRole: string;
-    isCurrentSpeaker: boolean;
-    onStartDebate: () => void;
-    onEndPresentation: () => void;
-    onToggleHand: () => void;
-    isMyHandRaised: boolean;
-}
-
-/** 라운드별 Action button */
-function RoundActions({
-                          roundType,
-                          myRole,
-                          isCurrentSpeaker,
-                          onStartDebate,
-                          onEndPresentation,
-                          onToggleHand,
-                          isMyHandRaised
-                      }: RoundActionsProps) {
-    return (
-        <>
-            {roundType === "PREPARATION" && myRole === 'HOST' && (
-                <ActionButton onClick={onStartDebate}>토론 시작하기</ActionButton>
-            )}
-            {roundType === "PRESENTATION" && (
-                <ActionButton
-                    disabled={!isCurrentSpeaker}
-                    onClick={onEndPresentation}
-                >
-                    발표 끝내기
-                </ActionButton>
-            )}
-            {roundType !== "PREPARATION" && <MicrophoneControl/>}
-            {
-                roundType === "FREE" && (
-                    <ActionButton
-                        onClick={onToggleHand}
-                        style={{
-                            backgroundColor: isMyHandRaised ? '#1976d2' : undefined,
-                            color: isMyHandRaised ? 'black' : undefined
-                        }}
-                    >
-                        <img
-                            src={raiseHandSvg}
-                            alt={isMyHandRaised ? "손내리기" : "손들기"}
-                            width={16.5}
-                            height={24}
-                            style={{
-                                filter: isMyHandRaised ? 'black' : undefined
-                            }}
-                        />
-                    </ActionButton>
-                )
-            }
-        </>
-    );
 }
 
 function DebatePageContent({debateId}: Props) {
